refactor(layout): consolidate Stack types and utils
- Merge BpkStack types into types.ts

- Inline Stack prop processing and remove redundant tokenUtils file
